Move handleNew() logic to own file
